RepeatBase in lib-builtin-effects
